@@ -1,15 +1,22 @@
 {
-  "name": "portfolio-website-main",
+  "name": "my_portfolio",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.4",
-    "@testing-library/user-event": "^13.5.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "5.0.0",
-    "web-vitals": "^2.1.4"
+    "@heroicons/react": "^1.0.6",
+    "@testing-library/jest-dom": "^5.14.1",
+    "@testing-library/react": "^13.0.0",
+    "@testing-library/user-event": "^13.2.1",
+    "framer-motion": "^7.2.0",
+    "nuka-carousel": "^5.3.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1",
+    "react-scroll": "^1.8.7",
+    "react-slider-swiper": "^2.0.6",
+    "swiper": "^8.3.2",
+    "web-vitals": "^2.1.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -36,13 +43,6 @@
     ]
   },
   "devDependencies": {
-    "@heroicons/react": "^1.0.6",
-    "autoprefixer": "^10.4.2",
-    "framer-motion": "^6.2.8",
-    "postcss": "^8.4.8",
-    "react-icons": "^4.3.1",
-    "react-scroll": "^1.8.6",
-    "swiper": "^8.0.7",
-    "tailwindcss": "^3.0.23"
+    "tailwindcss": "^3.1.8"
   }
 }
